# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LumoNAS API
  description: |
    LumoNAS management daemon API. Provides endpoints for system monitoring,
    storage management, Docker orchestration, networking, backups, and administration.
  version: 0.1.0
  contact:
    name: LumoNAS
    url: https://github.com/lumonas/lumonas
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Local development
...
paths:
...
  /files:
    get:
      tags: [Files]
      summary: List files
      operationId: listFiles
      parameters:
        - name: shareId
          in: query
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileEntry'
...
components:
  schemas:
...
    FileEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [dir, file]
        sizeBytes:
          type: integer
        modifiedAt:
          type: string
          format: date-time
...